requirements.txt
Flask==2.3.2
Gunicorn==21.2.0
requests==2.31.0
praw==7.8.1
python-dotenv==1.0.0
pytrends==4.8.0
